refactor: split types/index.ts into domain-specific files
Split the monolithic types/index.ts into year.ts, territory.ts, and
app-state.ts for better discoverability and consistency with pipeline.ts.
The barrel index.ts now only re-exports from the domain files.
